+++ src/components/postings/index.js
@@ -99,6 +99,11 @@ function Postings() {
     }
 
     const handleDelete = async (postingId) => {
+        const confirmDelete = window.confirm("Are you sure you want to delete this posting?");
+        if (!confirmDelete) {
+            return;
+        }
+
         const postingDoc = doc(db, 'postings', postingId);
 
         await deleteDoc(postingDoc);
@@ -128,8 +133,7 @@ function Postings() {
     }
 
     useEffect(() => {
-        console.log(tagsEnabled);
-        setPostings((prev) => 
+        setPostings((prev) =>
             prev.map((posting) => ({
                 ...posting,
                 isVisible: tagsEnabled.length === 0
@@ -141,24 +145,29 @@ function Postings() {
 
     return (
         <Card>
-            <Card.Body>
-                <Card.Header>
+            <Card.Body className='d-flex align-items-center flex-column'>
+                <Card.Header className='w-100'>
                     <NavBar role={role} />
                 </Card.Header>
-                <Card.Title>Job Postings</Card.Title>
-                {tags.map((tag) => (
-                    <Button
-                        key={tag}
-                        variant={tagsEnabled.includes(tag) ? "primary" : "secondary"}
-                        onClick={() => { toggleTag(tag) }}
-                    >
-                        {tag}
-                    </Button>
-                ))}
+                <Card.Title style={{ marginTop: '0.1em', fontSize: '2em' }}>Job Postings</Card.Title>
+                <div style={{ display: 'flex', flexDirection: 'row', marginBottom: '1em' }}>
+                    {tags.map((tag) => (
+                        <Button
+                            key={tag}
+                            variant={tagsEnabled.includes(tag) ? "success" : "secondary"}
+                            onClick={() => { toggleTag(tag) }}
+                            style={{ marginLeft: '0.5em', marginRight: '0.5em' }}
+                        >
+                            {tag}
+                        </Button>
+                    ))}
+                </div>
                 {role === "admin" && (
                     <>
-                        <Button onClick={() => toggleVisibility('approved')}>{hideApproved ? "Show Approved Posts" : "Hide Approved Posts"}</Button>
-                        <Button onClick={() => toggleVisibility('unapproved')}>{hideUnapproved ? "Show Unapproved Posts" : "Hide Unapproved Posts"}</Button>
+                        <div style={{ display: 'flex', flexDirection: 'row', marginBottom: '1em' }}>
+                            <Button style={{ marginLeft: '0.5em', marginRight: '0.5em' }} onClick={() => toggleVisibility('approved')}>{hideApproved ? "Show Approved Posts" : "Hide Approved Posts"}</Button>
+                            <Button style={{ marginLeft: '0.5em', marginRight: '0.5em' }} onClick={() => toggleVisibility('unapproved')}>{hideUnapproved ? "Show Unapproved Posts" : "Hide Unapproved Posts"}</Button>
+                        </div>
                     </>
                 )}
                 {loading ? (
@@ -169,44 +178,54 @@ function Postings() {
                         .map((posting) => (
                             (posting.status === "approved" && !hideApproved) ||
                                 (posting.status === "unapproved" && !hideUnapproved) ? (
-                                <Card key={posting.id}>
+                                <Card style={{ borderWidth: '2px', width: '50vw', marginBottom: '2em' }} key={posting.id}>
                                     <Card.Body>
-                                        <Card.Header className="d-flex">
-                                            <Card.Title>{posting.title}</Card.Title>
+                                        <Card.Header className="d-flex align-items-center justify-content-space-between">
+                                            {role !== 'admin' || role !==  'student' &&
+                                                <Card.Title className='w-100 text-center'><strong>{posting.title}</strong></Card.Title>
+                                            }
                                             {role === "student" && (
-                                                <Button
-                                                    disabled={applyLoading}
-                                                    onClick={() => handleApply(posting)}
-                                                >
-                                                    {applyLoading ? "Loading..." : posting.applicants.includes(user.uid) ? "Unapply" : "Apply"}
-                                                </Button>
+                                                <>
+                                                    <Card.Title style={{ flex: 1, textAlign: 'center'}} className='w-100 text-center'><strong>{posting.title}</strong></Card.Title>
+                                                    <Button
+                                                        disabled={applyLoading}
+                                                        onClick={() => handleApply(posting)}
+                                                    >
+                                                        {applyLoading ? "Loading..." : posting.applicants.includes(user.uid) ? "Unapply" : "Apply"}
+                                                    </Button>
+                                                </>
                                             )}
                                             {role === "admin" && (
                                                 <>
-                                                    <Button onClick={() => handleApprove(posting)}>{posting.status === "approved" ? "unapprove" : "approve"}</Button>
-                                                    <Button onClick={() => handleDelete(posting.id)} variant="danger">Delete Posting</Button>
+                                                    <Button style={{ margin: '0', width: '250px' }} onClick={() => handleDelete(posting.id)} variant="danger">Remove Posting</Button>
+                                                    <Card.Title className='w-100 text-center'><strong>{posting.title}</strong></Card.Title>
+                                                    <Button onClick={() => handleApprove(posting)}>{posting.status === "approved" ? "Unapprove" : "Approve"}</Button>
                                                 </>
                                             )}
                                         </Card.Header>
                                         <Card.Text>
-                                            {posting.selectedTags?.length > 0
-                                                ? posting.selectedTags.map((tag) => (
-                                                    <Button key={tag} variant="info" disabled>
-                                                        {tag}
-                                                    </Button>
-                                                ))
-                                                : "No tags"}{" "}
-                                            <br />
-                                            Location: {posting.location} <br />
-                                            Address: {posting.address} <br />
-                                            {role === "admin" && <>Status: {posting.status} <br /></>}
+                                            <div style={{ width: '100%', display: 'flex', flexDirection: 'row', marginBottom: '0.5em', marginTop: '1em' }}>
+                                                {posting.selectedTags?.length > 0
+                                                    ? posting.selectedTags.map((tag) => (
+                                                        <Button style={{ marginRight: '1em' }} key={tag} variant="success" disabled>
+                                                            {tag}
+                                                        </Button>
+                                                    ))
+                                                    : "No tags"}{" "}
+                                            </div>
+                                            <div>
+                                                <strong>Location:</strong> {posting.location} <br />
+                                                <strong>Address:</strong> {posting.address} <br />
+                                            </div>
+                                            {role === "admin" && <><strong>Status:</strong> {posting.status} <br /></>}
                                             <div style={{ maxHeight: '300px', overflowY: 'auto' }}>
-                                                <h5>Description:</h5>
+                                                <Card.Text style={{ marginBottom: '0.5em' }}><strong>Job Description:</strong></Card.Text>
                                                 <ReactQuill
                                                     value={posting.jobDescription}
                                                     readOnly
                                                     theme="snow"
                                                     modules={{ toolbar: false }}
+                                                    style={{ marginBottom: '0.5em' }}
                                                 />
                                             </div>
                                             Date Published: {handleDate(posting.datePublished)}
@@ -218,7 +237,6 @@ function Postings() {
                 ) : (
                     <Card.Text>No postings found</Card.Text>
                 )}
-
             </Card.Body>
         </Card>
     )
